requirements.txt
bcrypt==4.3.0
boto3==1.37.12
google-generativeai==0.8.5
python-dotenv==1.1.0
openpyxl==3.1.5
pandas==2.2.0
streamlit==1.42.2
supabase==2.28.0
pillow==11.1.0
psycopg2-binary==2.9.10
